refactor(uploads): Keep model ids out of the client bundle
Declaring consumers next to the accepted types put both chat model ids
and the PDF adapter into a browser chunk, which the client has no use
for: it only needs the types and the size cap.

The consumer map moves to its own module that browser code never
imports. The invariant is unchanged — the test reads both halves.

See also: #782
