+++ src/lib/convex/_generated/api.d.ts
@@ -77,10 +77,13 @@ import type * as support_handoff from "../support/handoff.js";
 import type * as support_messageListing from "../support/messageListing.js";
 import type * as support_messages from "../support/messages.js";
 import type * as support_migration from "../support/migration.js";
+import type * as support_notificationPreferences from "../support/notificationPreferences.js";
 import type * as support_ownership from "../support/ownership.js";
 import type * as support_promptStore from "../support/promptStore.js";
 import type * as support_rateLimit from "../support/rateLimit.js";
 import type * as support_supportThreadFields from "../support/supportThreadFields.js";
+import type * as support_threadLifecycle from "../support/threadLifecycle.js";
+import type * as support_threadMaintenance from "../support/threadMaintenance.js";
 import type * as support_threads from "../support/threads.js";
 import type * as support_tools_handoff from "../support/tools/handoff.js";
 import type * as support_types from "../support/types.js";
@@ -165,10 +168,13 @@ declare const fullApi: ApiFromModules<{
   "support/messageListing": typeof support_messageListing;
   "support/messages": typeof support_messages;
   "support/migration": typeof support_migration;
+  "support/notificationPreferences": typeof support_notificationPreferences;
   "support/ownership": typeof support_ownership;
   "support/promptStore": typeof support_promptStore;
   "support/rateLimit": typeof support_rateLimit;
   "support/supportThreadFields": typeof support_supportThreadFields;
+  "support/threadLifecycle": typeof support_threadLifecycle;
+  "support/threadMaintenance": typeof support_threadMaintenance;
   "support/threads": typeof support_threads;
   "support/tools/handoff": typeof support_tools_handoff;
   "support/types": typeof support_types;
