+++ src/lib/uploads/profiles.ts
@@ -6,28 +6,11 @@
  * validator cannot drift apart. They used to be three independent literals
  * (see #782), and the profile-image surface had already drifted.
  *
- * A profile also names its consumer, which is what makes the allowlist
- * checkable: `profiles.test.ts` asserts that a surface only accepts media its
- * consumer can handle. The consumer is not always a model — an avatar is only
- * ever rendered — so it is modelled as either a model or a plain pipeline.
- *
- * Imported by both browser and Convex code, so this module stays free of
- * Svelte, browser APIs and `convex/values`.
+ * Which model or pipeline reads the files lives in `consumers.ts` instead, so
+ * importing a profile in browser code does not pull model ids into the client
+ * bundle. Imported by both browser and Convex code, so this module stays free
+ * of Svelte, browser APIs and `convex/values`.
  */
-
-import { AI_CHAT_MODEL_ID, SUPPORT_MODEL_ID } from '../convex/utils/chatModel';
-import { mediaCategoryOf, MODEL_INPUT_MODALITIES } from './model-capabilities';
-
-/**
- * Who ends up processing the bytes.
- *
- * `adapters` records a media type the model itself cannot read but something
- * in the request path converts on its behalf. The value is the reason, not a
- * boolean, so the exception has to be spelled out to be granted.
- */
-export type Consumer =
-	| { kind: 'model'; modelId: string; adapters?: Readonly<Record<string, string>> }
-	| { kind: 'pipeline'; id: string; handles: readonly string[] };
 
 export type UploadProfile = {
 	/** Extension to mime type. The source of truth both derivations read. */
@@ -35,22 +18,7 @@ export type UploadProfile = {
 	maxBytes: number;
 	maxBytesLabel: string;
 	maxFiles: number;
-	/**
-	 * Everything that ends up reading these files. A surface can feed more than
-	 * one: chat attachments reach both the assistant and the support agent, and
-	 * `chatModel.ts` keeps those model ids apart precisely so they can diverge.
-	 * The profile is only valid if *every* consumer can handle what it accepts.
-	 */
-	consumers: readonly Consumer[];
 };
-
-/**
- * PDFs reach the chat models through OpenRouter's file parser, not through the
- * models themselves, which report no `file` modality. The engine is pinned at
- * the call site in `aiUsage/capture.ts`; leaving it unset bills Mistral OCR per
- * page (#781).
- */
-const PDF_VIA_OPENROUTER = { 'application/pdf': 'openrouter:file-parser:cloudflare-ai' } as const;
 
 export const UPLOAD_PROFILES = {
 	/** Attachments on the AI chat and support surfaces. */
@@ -67,11 +35,7 @@ export const UPLOAD_PROFILES = {
 		},
 		maxBytes: 5 * 1024 * 1024,
 		maxBytesLabel: '5MB',
-		maxFiles: 6,
-		consumers: [
-			{ kind: 'model', modelId: AI_CHAT_MODEL_ID, adapters: PDF_VIA_OPENROUTER },
-			{ kind: 'model', modelId: SUPPORT_MODEL_ID, adapters: PDF_VIA_OPENROUTER }
-		]
+		maxFiles: 6
 	},
 	/**
 	 * The account avatar. No model involved: the file is downscaled and shown.
@@ -91,8 +55,7 @@ export const UPLOAD_PROFILES = {
 		},
 		maxBytes: 2 * 1024 * 1024,
 		maxBytesLabel: '2MB',
-		maxFiles: 1,
-		consumers: [{ kind: 'pipeline', id: 'avatar-render', handles: ['image'] }]
+		maxFiles: 1
 	}
 } as const satisfies Record<string, UploadProfile>;
 
@@ -112,30 +75,4 @@ export function acceptAttribute(profile: UploadProfile): string {
 export function acceptsMimeType(profile: UploadProfile, mimeType: string): boolean {
 	const essence = mimeType.split(';')[0]!.trim().toLowerCase();
 	return allowedMimeTypes(profile).includes(essence);
-}
-
-/**
- * Whether a consumer can take one specific mime type.
- *
- * Resolved per mime type rather than per category on purpose: an adapter is
- * declared for `application/pdf`, and widening that to the whole `file`
- * category would silently also permit `.docx` and every other document type.
- *
- * `unknown-model` is its own answer so a model configured without a capability
- * entry fails the test instead of quietly passing it.
- */
-export type ConsumeVerdict = 'native' | 'adapter' | 'unsupported' | 'unknown-model';
-
-export function canConsume(consumer: Consumer, mimeType: string): ConsumeVerdict {
-	const essence = mimeType.split(';')[0]!.trim().toLowerCase();
-	const category = mediaCategoryOf(essence);
-
-	if (consumer.kind === 'pipeline') {
-		return consumer.handles.includes(category) ? 'native' : 'unsupported';
-	}
-
-	const modalities = MODEL_INPUT_MODALITIES[consumer.modelId];
-	if (!modalities) return 'unknown-model';
-	if (modalities.includes(category)) return 'native';
-	return consumer.adapters?.[essence] ? 'adapter' : 'unsupported';
 }
